# serveur-REST-depenses-v0.2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Serveur des dépenses partagées",
        "description": "Utilisation d'un serveur REST",
        "contact": {
            "email": "fred.hemery@univ-artois.fr"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Le serveur pour les dépenses partagées"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Membres"
                ],
                "summary": "Enregistrement d'un membre",
                "description": "Enregistrement d'un membre",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prenom": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nom": "Duchmol",
                                    "prenom": "Robert",
                                    "email": "robert.duchmol@domain.com",
                                    "password": "secret"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "SQL Error or Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Membres"
                ],
                "summary": "connexion d'un membre",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "robert.duchmol@domain.com",
                                    "password": "secret"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success: operation Successfully"
                    },
                    "401": {
                        "description": "Refused: Invalid Credentials"
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Membres"
                ],
                "summary": "déconnexion d'un membre",
                "operationId": "App\\Http\\Controllers\\API\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "Success: operation Successfully"
                    },
                    "401": {
                        "description": "Refused: Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
...
}